network: fixes to public address support
1. requestTracker changes:
  - Remove http.Request.RemoteAddr overwriting in request tracker
  - Remove http.Request from request tracker
  - Add a new remoteAddresss() method that returns a most suitable address

2. public addr changes:
  - no otherPublicAddr outside of requestTracker
  - new remoteAddresss() method provides most meaningful address for incoming requests

Rationale:
There is a chain http.Handler's in wsNetwork: request tracker and wsNetwork itself.
Tracked request is created/updated in the first ServeHTTP and used to save a pointer to http.Request object.
This request object is actually gets copied in mux.ServeHTTP that calls downstream wsNetwork's ServeHTTP making
the request non-usable in the main wsNetwork's ServeHTTP so removed.

otherPublicAddr is an address reported by a peer via our custom X-Algorand-Location header that the only
correct value for wsPeer.rootURL (host:port accepting connections) but cannot be trusted.
Unconditionally rewriting it with remoteAddr obtained from a tcp connection is too much and a new
remoteAddresss() method attempts to provide the most meaningful trusted value.

Fixes #5713