AVM: Add a lot of type annotations to opcodes
This allows the assembler to complain about clear mistakes, but maybe
more importantly, it improves the specs/documentation.  We can now
document all the required byte lengths, even if they are strange, but
using "b{33}" in `proto()` call to require a 33 byte input.

But we also just missed a bunch of opportunities to annotate opcodes
in the past.  I tried to consider ever single "b" in opcodes.go.

things to consider in review:

1) This should not change any behavior of assembled code. If it does,
that's a bug.

2) This _could_ make something that used to assemble stop
assembling. Probably, that means the author has a bug. If they really
want to call these opcodes with the wrong sized inputs, the can, using
`#pragma typetrack false`

3) Should we eliminate the existing single character names in proto()
that are only used to indicate a byte length?  For example
```
{0x84, "ed25519verify_bare", opEd25519VerifyBare, proto("b63:T"), 7, costly(1900)},
```
could now be written
```
{0x84, "ed25519verify_bare", opEd25519VerifyBare, proto("b{64}b{32}:T"), 7, costly(1900)},
```

Perhaps that is clearer.