Record the block payout explicitly
Add ProposerPayout to the block header to make it very clear how much
will be moved from the feesink to the proposer.

It is tentatively recorded during block evaluation, but agreement can
zero out the payment because it is only in the agreement code that the
actual proposer is known - so it can be checked for eligibility.