Leave lookback at 320 for suspension test
Part of the test requires the suspended node to re-propose after
suispension. That can only happen when the balance lookback is
considerably bigger than the suspension interval (here, 5*10==50)

Fortunately, this test doesn't have to wait for lookback, so the
benefit of short filter timeout is all we need for a quick test.