tests: fix flushing in TestVotersReloadFromDiskAfterOneStateProofCommitted
The test in question uses a combination of automatic commits and a manual commit triggering.
Stopping commitSyncer in the middle of the test creates a race condition between
enqueueing and consuming - it is likely a task enqueued after commitSyncer is done draining
and exiting. This leaves us with one element in deferredCommits (that is OK) and one extra
accountsWriting wait group element (not OK on close if commitSyncer is active).
Given a later reloadLedger call leads to restarting commitSyncer, re-initalization the queue
but preserving accountsWriting value that is one off and drives infinite wait on closing.

The fix is just updates the test to prevent the race above.
It is also possible to reset accountsWriting in trackerRegistry.initialize but it looks useful
to catch closing errors in our unit tests exercising various closing scenarios.