ensure resource population is deterministic
The inital reason for desiring this was because of the non-deterministic
behavior of maps made testing difficult. When thinking about it more, I
realized that having deterministic population will also improve the
developer experience since the same txn group will always get back the
same populated resources. This change, however, does expose a potential
problem inherit to resource population: order matters. As seen in the
modified test, the determinstic order results in an extra resource in
the extra resource array. In the future steps could be taken to try to
improve the efficiency, but unless we go over every permutation of
resource ordering there will always be cases where algorithmic
population will not result in the most efficient resource packing.