Refactor some stateless checks on transactions
Some cleanups I ran into while doing tx.Access.  Wanted to make this
PR first because that PR risks being big anyway.

transaction.WellFormed() was huge. Move each transaction's
implementation into the file that defined the transaction type's
field.

transaction.Alive() is better thought of as a method on block.  It's
the block that decides if the transaction is Alive. By moving it
there, we remove an interface that was only introduced to fix the
circular dependency caused by putting in the "wrong" place.