Setup lobgoal interfaces to enable tx.Access packing
Interfaces we previously designed to take all the foreign arrays,
ready for packing into the transaction. Now they receive a "RefBundle"
of things the tx needs access to, so it should be possible to change
that to use tx.Access.  Committing now as a checkpoint that should not
have changed any functionality.