Add goal switches to populate access list transactions.
This required changes to libgoal interfaces that we previously
designed to take all the foreign arrays, ready for packing into the
transaction. Now they receive a "RefBundle" of things the tx needs
access to, so it should be possible to change that to use tx.Access.

Still needs final e2e subs tests using goal.