Allow access of boxes for apps in made in the same group
We know they are empty, so we need not worry about exceeded read
quota.  We will still treat them as dirty, so we need not worry about
exceed write quota.

We bound the number of boxes that can be accessed this way by the
number of empty box refs.  This means that the _number_ of boxes
written is bounded, not just the sum of the lengths.