simulate support for newly created apps creating boxes
This includes some additional bugs fixes.  Previously, `simulate`
would report a boxref that included a newly created app's ID when used
with AllowUnnamedResources.  That is, in some sense, correct, but
useless, since a follow-on transaction that used such a box ref would
surely have the wrong app id, and would still fail.

Now, such a box ref will be converted to an empty ref if
EnableUnnamedBoxAccessInNewApps is true.

I considered always converting, but there are corner cases in which a
very clever user of simulate could correctly use the old result.
After the change, the new result is always strictly easier to use.