Move all the allobounds globals into `bounds` package
By moving bounds into their own package, it is both
clearer (config.MaxTxGroupSize was never a _config_ value, really, it
was a derived bound, based on config values.) and allows the `basic`
to import only these limits, with importing `config`.  In turn, that
will mean that we can import `basic` into `config` so
`ConsensusParams` can have types in it like `basics.Round`.