app_params_set, new box params, and "foreign" box opcodes
This PR introduces `app_params_set` as a way for apps to set some of
their own parameters. As of this PR, two flags are introduced.
ForeignBoxReads makes all of the calling app's boxes "public" in the
sense that they can be read by other apps, on-chain.  FamilyBoxAccess,
makes all of the boxes available for read AND write, but only to apps
with the same creator as the calling app.

All box opcodes have a new version which accepts the appID for the box
being manipulated.  For example `box_len` has a new version called
`app_box_len` that takes an additional argument, the appID of the box
whose length is being queried.  Perhaps these should be called
`box_len_ex` by analogy to `app_global_get_ex` which has a similar
relationship to `app_global_get`.

Of course, the box being accessed must be _available_ by the usual
meaning of _available_. (Listed in tx.Boxes or tx.Access, in one of
the transactions of the group.)