Make counts and maxCounts simple values
It's makes it clearer that there is no sharing among the
storageDeltas. There was no sharing, because each pointed to object
was freshly allocated, but using a pointer obscured that fact (and
performed a wasted allocation.

This change will make it easier (or at least clearer) to make changes
to `maxCounts` if we allow transactions that update an app's global
schema.