AVM: Allow 0 as app in LocalRef, specifying current app
A localref inside of tx.Access ought to be able to use 0 for "current
app". It was an oversight to prohibit it.  This PR allows it, but is
consensus gated.

This was a slightly bigger that chnage than expected (because the
access list must now be scanned at app create time for localrefs woth
0 apps, like boxes were). So a few tests are needed before this is
reviewable.

Also removes two consensus flags that can be assumed true now because
they are strictly more lenient.