Chore: Eliminate BlockEvaluator.Transaction
The existence of a public BlockEvaluator.Transaction perpetuates the
myth that it's reasonable to evaluate transactions one at a time.  No
"real" code used this method, only tests, which actually meant
"construct a one element Group and evaluate it.  To eliminate it, and
keep the test callers from looking cluttered, TransactionGroup() was
made variadic, and WithAD() was added as a convenient way to construct
a SignedTransactionWithAD from SignedTransaction.

This means that called to TransactionGroup(x) become
TransactionGroup(x...), but note that this does not allocate or copy.
Go Spec says this about variadic functions:

```
If the final argument is assignable to a slice type []T and is followed by ..., it is passed unchanged as the value for a ...T parameter. In this case no new slice is created.

Given the slice s and call

s := []string{"James", "Jasmine"}
Greeting("goodbye:", s...)
within Greeting, who will have the same value as s with the same underlying array.
```